requirements.txt
pywin32==306
pystray==0.19.3
Pillow==10.0.0
pyinstaller==5.13.0